backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
langchain==0.1.20
langchain-openai==0.1.6
langchain-community==0.0.38
langchain-core==0.1.52
chromadb==0.4.18
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
openai>=1.24.0,<2.0.0
pydantic==2.5.2
python-multipart==0.0.6
httpx==0.25.2
packaging>=23.2,<24.0

